+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- [hardis:org:file:export](https://sfdx-hardis.cloudity.com/hardis/org/files/export/): Provide record Ids in the logs
+- Update labels of all report files for UI buttons 
+
 ## [6.4.2] 2025-09-10
 
 - Add additional dependencies in Ubuntu Dockerfile to allow mermaid-cli and chrome to run natively from the image.